birdtrends: Estimate Trends Based On Modelled Annual Indices Or Related Inputs

This package provides a flexible framework to generate trends, based on actual or modeled annual indices data. This package was designed to undertake assessment of 2016 Partners In Flight Landbird Conservation objectives.

See also

Author

Maintainer: Genevieve Perkins genevieve.c.perkins@gmail.com (ORCID)

Authors:

Adam C. Smith adam.smith2@canada.ca